more work on eager / lazy separation
