+++ src/main/java/com/aol/simple/react/stream/traits/FutureStream.java
@@ -1,16 +1,22 @@
 package com.aol.simple.react.stream.traits;
 
+import static org.jooq.lambda.tuple.Tuple.tuple;
+
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
+import java.util.ListIterator;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Random;
 import java.util.Spliterator;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.locks.LockSupport;
 import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
 import java.util.function.BiPredicate;
@@ -23,21 +29,48 @@ import java.util.function.Supplier;
 import java.util.function.ToDoubleFunction;
 import java.util.function.ToIntFunction;
 import java.util.function.ToLongFunction;
+import java.util.stream.Collector;
+import java.util.stream.Collectors;
 import java.util.stream.DoubleStream;
 import java.util.stream.IntStream;
 import java.util.stream.LongStream;
 import java.util.stream.Stream;
 
+import lombok.AllArgsConstructor;
+
 import org.jooq.lambda.Seq;
+import org.jooq.lambda.tuple.Tuple;
 import org.jooq.lambda.tuple.Tuple2;
 
 import com.aol.simple.react.async.Queue;
+import com.aol.simple.react.async.Queue.ClosedQueueException;
+import com.aol.simple.react.async.Queue.QueueReader;
+import com.aol.simple.react.async.Queue.QueueTimeoutException;
+import com.aol.simple.react.async.factories.QueueFactories;
+import com.aol.simple.react.async.factories.QueueFactory;
 import com.aol.simple.react.async.subscription.Continueable;
+import com.aol.simple.react.exceptions.ExceptionSoftener;
+import com.aol.simple.react.exceptions.SimpleReactFailedStageException;
 import com.aol.simple.react.stream.CloseableIterator;
+import com.aol.simple.react.stream.traits.future.operators.ToLazyCollection;
+import com.aol.simple.react.stream.traits.operators.BatchBySize;
+import com.aol.simple.react.stream.traits.operators.BatchByTime;
+import com.aol.simple.react.stream.traits.operators.BatchByTimeAndSize;
+import com.aol.simple.react.stream.traits.operators.Debounce;
+import com.aol.simple.react.stream.traits.operators.SlidingWindow;
+import com.aol.simple.react.util.SimpleTimer;
 
-public interface FutureStream<U> extends Seq<U> {
+public interface FutureStream<U> extends Seq<U>, SimpleReactStreamInterface<U>, ToQueue<U> {
 
+	static final ExceptionSoftener softener = ExceptionSoftener.singleton.factory
+			.getInstance();
 	
+	static <T,R> Function<FutureStream<T>,FutureStream<R>> lift(Function<T,R> fn){
+		return fs -> fs.map(v->fn.apply(v));
+	}
+	static <T1,T2,R> BiFunction<FutureStream<T1>,FutureStream<T2>,FutureStream<R>> lift2(BiFunction<T1,T2,R> fn){
+		return (fs1,fs2) -> fs1.flatMap( v1-> (FutureStream)fs2.map(v2->fn.apply(v1,v2)));
+	}
 	
 	/**
 	 * 
@@ -53,7 +86,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * 			Collection will not be populated until methods are accessed.
 	 * 			(Eager streams may populate a LazyColleciton in the background however)
 	 */
-	Collection<U> toLazyCollection();
+	default Collection<U> toLazyCollection(){
+		return ToLazyCollection.toLazyCollection(this.iterator());
+	}
 	/**
 	 * Create a LazyCollection access to which is synchronized
 	 * <pre>
@@ -68,7 +103,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * 			Collection will not be populated until methods are accessed.
 	 * 			(Eager streams may populate a LazyColleciton in the background however)
 	 */
-	Collection<U> toConcurrentLazyCollection();
+	default Collection<U> toConcurrentLazyCollection(){
+		return ToLazyCollection.toConcurrentLazyCollection(this.iterator());
+	}
 	/**
 	 * <pre>
 	 * {@code 
@@ -81,56 +118,103 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @return single element from this Stream if it is a single element Stream
 	 * 			otherwise throw an UnsupportedOperationException
 	 */
-	U single();
+	default U single(){
+		 List<U> l= toList(); 
+		 if(l.size()==1){ 
+			 return l.get(l.size()-1); 
+			 }
+		throw new UnsupportedOperationException("single only works for Streams with a single value");
+	}
 	/**
 	 * Zip two Streams, zipping against the underlying futures of this stream
 	 * 
 	 * @param other
 	 * @return
 	 */
-	<R> FutureStream<Tuple2<U,R>> zipFutures(Stream<R> other);
+	 <R> FutureStream<Tuple2<U,R>> zipFutures(Stream<R> other);/** {
+		if(other instanceof FutureStream)
+			return zipFutures((FutureStream)other);
+		Seq seq = Seq.seq(getLastActive().stream()).zip(Seq.seq(other));
+		Seq<Tuple2<CompletableFuture<U>,R>> withType = (Seq<Tuple2<CompletableFuture<U>,R>>)seq;
+		Stream futureStream = fromStreamOfFutures((Stream)withType.map(t ->t.v1.thenApply(v -> Tuple.tuple(t.v1.join(),t.v2)))
+				);
+
+		
+		return (FutureStream<Tuple2<U,R>>)futureStream;
+
+	
+	}**/
+	
 	/* 
 	 * @see org.jooq.lambda.Seq#crossJoin(java.util.stream.Stream)
 	 */
 	@Override
-	 <T> FutureStream<Tuple2<U, T>> crossJoin(Stream<T> other);
+	default <T> FutureStream<Tuple2<U, T>> crossJoin(Stream<T> other) {
+	        return fromStream(Seq.crossJoin(this, other));
+	}
+	
+	
 	/**
      * Produce this stream, or an alternative stream from the
      * {@code value}, in case this stream is empty.
      */
 	@Override
-	FutureStream<U> onEmpty(U value);
+	default FutureStream<U> onEmpty(U value){
+		return fromStream(Seq.super.onEmpty(value));
+	}
 	/**
      * Produce this stream, or an alternative stream from the
      * {@code supplier}, in case this stream is empty.
      */
 	@Override
-	FutureStream<U> onEmptyGet(Supplier<U> supplier);
+	default FutureStream<U> onEmptyGet(Supplier<U> supplier){
+		return fromStream(Seq.super.onEmptyGet(supplier));
+	}
 	 /**
      * Produce this stream, or an alternative stream from the
      * {@code supplier}, in case this stream is empty.
      */
 	@Override
-	<X extends Throwable> FutureStream<U> onEmptyThrow(Supplier<X> supplier);
+	default <X extends Throwable> FutureStream<U> onEmptyThrow(Supplier<X> supplier) {
+			return fromStream(Seq.super.onEmptyThrow(supplier));
+	}
     
 	/* 
 	 * @see org.jooq.lambda.Seq#innerJoin(java.util.stream.Stream, java.util.function.BiPredicate)
 	 */
 	@Override
-    <T> FutureStream<Tuple2<U, T>> innerJoin(Stream<T> other, BiPredicate<U, T> predicate);
+    default <T> FutureStream<Tuple2<U, T>> innerJoin(Stream<T> other, BiPredicate<U, T> predicate) {
+
+       
+        RepeatableStream<T> s = new RepeatableStream<>(ToLazyCollection.toLazyCollection(other.iterator()));
+
+        return flatMap(t -> s.stream()
+                           .filter(u -> predicate.test(t, u))
+                           .map(u -> tuple(t, u)));
+    }
 
     
     /* 
      * @see org.jooq.lambda.Seq#leftOuterJoin(java.util.stream.Stream, java.util.function.BiPredicate)
      */
 	@Override
-     <T> FutureStream<Tuple2<U, T>> leftOuterJoin(Stream<T> other, BiPredicate<U, T> predicate);
+    default <T> FutureStream<Tuple2<U, T>> leftOuterJoin(Stream<T> other, BiPredicate<U, T> predicate) {
+
+    	 RepeatableStream<T> s = new RepeatableStream<>(ToLazyCollection.toLazyCollection(other.iterator()));
+
+        return flatMap(t -> Seq.seq(s.stream())
+                           .filter(u -> predicate.test(t, u))
+                           .onEmpty(null)
+                           .map(u -> tuple(t, u)));
+    }
 
 	/* 
 	 * @see org.jooq.lambda.Seq#rightOuterJoin(java.util.stream.Stream, java.util.function.BiPredicate)
 	 */
 	@Override
-    <T> FutureStream<Tuple2<U, T>> rightOuterJoin(Stream<T> other, BiPredicate<U, T> predicate);
+    default <T> FutureStream<Tuple2<U, T>> rightOuterJoin(Stream<T> other, BiPredicate<U, T> predicate) {
+        return fromStream(Seq.super.rightOuterJoin(other, predicate));
+    }
    
 	
 	
@@ -147,15 +231,93 @@ public interface FutureStream<U> extends Seq<U> {
 	 */
 
 	<R> FutureStream<Tuple2<U,R>> zipFutures(FutureStream<R> other);
+	/**{
+		Seq seq = Seq.seq(getLastActive().stream()).zip(Seq.seq(other.getLastActive().stream()));
+		Seq<Tuple2<CompletableFuture<U>,CompletableFuture<R>>> withType = (Seq<Tuple2<CompletableFuture<U>,CompletableFuture<R>>>)seq;
+		Stream futureStream =  fromStreamOfFutures((Stream)withType.map(t ->CompletableFuture.allOf(t.v1,t.v2).thenApply(v -> Tuple.tuple(t.v1.join(),t.v2.join())))
+				);
+		
+		return (FutureStream<Tuple2<U,R>>)futureStream;
+		
+
+	}**/
 	
 	/**
 	 * @return an Iterator that chunks all completed elements from this stream since last it.next() call into a collection
 	 */
-	Iterator<Collection<U>> chunkLastReadIterator();
+	default Iterator<Collection<U>> chunkLastReadIterator(){
+		
+		Queue.QueueReader reader =  new Queue.QueueReader(this.withQueueFactory(QueueFactories.unboundedQueue())
+																	.toQueue(q->q.withTimeout(100)
+																	.withTimeUnit(TimeUnit.MICROSECONDS))
+																	,null);
+		class Chunker implements Iterator<Collection<U>> {
+			volatile boolean open =true;
+			@Override
+			public boolean hasNext() {
+
+				return open == true && reader.isOpen();
+			}
+
+			@Override
+			public Collection<U> next() {
+				
+				while(hasNext()){
+					try{
+						return reader.drainToOrBlock();
+					}catch(ClosedQueueException e){
+						open =false;
+						return new ArrayList<>();
+					}catch(QueueTimeoutException e){
+						LockSupport.parkNanos(0l);
+					}
+				}
+				return new ArrayList<>();
+				
+
+			}
+		}
+		return new Chunker();
+	}
 	/**
 	 * @return a Stream that batches all completed elements from this stream since last read attempt into a collection
 	 */
-	 FutureStream<Collection<U>> chunkSinceLastRead();
+	default FutureStream<Collection<U>> chunkSinceLastRead(){
+		Queue queue = this.withQueueFactory(QueueFactories.unboundedQueue()).toQueue();
+		Queue.QueueReader reader =  new Queue.QueueReader(queue,null);
+		class Chunker implements Iterator<Collection<U>> {
+			
+			@Override
+			public boolean hasNext() {
+
+				return reader.isOpen();
+			}
+
+			@Override
+			public Collection<U> next() {
+				return reader.drainToOrBlock();
+
+			}
+		}
+		Chunker chunker = new Chunker();
+		Function<Supplier<U>, Supplier<Collection<U>>> fn = s -> {
+			return () -> {
+				
+				try {
+					return chunker.next();
+				} catch (ClosedQueueException e) {
+					
+					throw new ClosedQueueException();
+				}
+				
+			};
+		};
+		return fromStream(queue.streamBatchNoTimeout(getSubscription(), fn));
+		
+		
+	}
+	FutureStream<U> withQueueFactory(
+			QueueFactory<U> qf);
 	/**
 	 * Break a stream into multiple Streams based of some characteristic of the elements of the Stream
 	 * 
@@ -171,13 +333,23 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param sharder Function to split split incoming elements into shards
 	 * @return Map of new sharded Streams
 	 */
-	<K> Map<K, ? extends FutureStream<U>> shard(
-			Map<K, Queue<U>> shards, Function<U, K> sharder) ;
+	default <K> Map<K, ? extends FutureStream<U>> shard(
+			Map<K, Queue<U>> shards, Function<U, K> sharder) {
+		toQueue(shards, sharder);
+		return shards
+				.entrySet()
+				.stream()
+				.collect(
+						Collectors.toMap(e -> e.getKey(), e -> fromStream(e
+								.getValue().stream(getSubscription()))));
+	}
 	
 	/**
 	 * Cancel the CompletableFutures in this stage of the stream
 	 */
-	void cancel();
+	default void cancel(){
+		this.streamCompletableFutures().forEach(next-> next.cancel(true));
+	}
 	
 	/**
 	 * Batch the elements in the Stream by a combination of Size and Time
@@ -187,7 +359,13 @@ public interface FutureStream<U> extends Seq<U> {
 	 * 
 	 * @return
 	 */
-	FutureStream<List<U>> batchBySizeAndTime(int size,long time, TimeUnit unit);
+	default FutureStream<List<U>> batchBySizeAndTime(int size,long time, TimeUnit unit) { 
+	 
+	    Queue<U> queue = toQueue();
+	    Function<BiFunction<Long,TimeUnit,U>, Supplier<Collection<U>>> fn = new BatchByTimeAndSize<>(size,time,unit,()->new ArrayList<>());
+	    return (FutureStream)fromStream(queue.streamBatch(getSubscription(), (Function)fn)).filter(c->!((Collection)c).isEmpty());
+	}
+	
 	/**
 	 * Batch the elements in the Stream by a combination of Size and Time
 	 * If batch exceeds max size it will be split
@@ -196,7 +374,12 @@ public interface FutureStream<U> extends Seq<U> {
 	 * 
 	 * @return
 	 */
-	<C extends Collection<U>> FutureStream<C> batchBySizeAndTime(int size,long time, TimeUnit unit, Supplier<C> factory);
+	default <C extends Collection<U>> FutureStream<C> batchBySizeAndTime(int size,long time, TimeUnit unit, Supplier<C> factory) { 
+	 
+	    Queue<U> queue = toQueue();
+	    Function<BiFunction<Long,TimeUnit,U>, Supplier<Collection<U>>> fn = new BatchByTimeAndSize(size,time,unit,factory);
+	    return (FutureStream)fromStream(queue.streamBatch(getSubscription(), (Function)fn));
+	}
 	/**
 	 * 
 	 * Batch the elements in this stream into Lists of specified size
@@ -204,14 +387,21 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param size Size of lists elements should be batched into
 	 * @return Stream of Lists
 	 */
-	FutureStream<List<U>> batchBySize(int size);
+	default FutureStream<List<U>> batchBySize(int size) {
+		Queue queue = toQueue();
+		Function<Supplier<U>, Supplier<List<U>>> fn = new BatchBySize(size,this.getSubscription(),queue,()->new ArrayList<>());
+		return fromStream(queue.streamBatchNoTimeout(getSubscription(), fn));
+	}
 	/**
 	 * Batch elements into a Stream of collections with user defined function
 	 * @param fn Function takes a supplier, which can be used repeatedly to get the next value from the Stream. If there are no more values, a ClosedQueueException will be thrown.
 	 *           This function should return a Supplier which creates a collection of the batched values
 	 * @return Stream of batched values
 	 */
-	<C extends Collection<U>>FutureStream<C> batch(Function<Supplier<U>, Supplier<C>> fn);
+	default <C extends Collection<U>>FutureStream<C> batch(Function<Supplier<U>, Supplier<C>> fn){
+		Queue queue = toQueue();
+		return fromStream(queue.streamBatchNoTimeout(getSubscription(), fn));
+	}
 	
 	/**
 	 * Batch the elements in this stream into Collections of specified size
@@ -221,7 +411,11 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param supplier Create the batch holding collection
 	 * @return Stream of Collections
 	 */
-	<C extends Collection<U>>FutureStream<C> batchBySize(int size, Supplier<C> supplier);
+	default <C extends Collection<U>>FutureStream<C> batchBySize(int size, Supplier<C> supplier) {
+		Queue queue = toQueue();
+		Function<Supplier<U>, Supplier<Collection<U>>> fn = new BatchBySize(size,this.getSubscription(),queue,supplier);
+		return fromStream(queue.streamBatchNoTimeout(getSubscription(), fn));
+	}
 	/**
 	 * Introduce a random delay between events in a stream
 	 * Can be used to prevent behaviour synchronizing within a system
@@ -229,7 +423,35 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param jitterInNanos Max number of nanos for jitter (random number less than this will be selected)/
 	 * @return Next stage in Stream with jitter applied
 	 */
-	FutureStream<U> jitter(long jitterInNanos);
+	default FutureStream<U> jitter(long jitterInNanos){
+		Queue queue = toQueue();
+		Random r = new Random();
+		Function<Supplier<U>, Supplier<U>> fn = s -> {
+			return () -> {
+				SimpleTimer timer = new SimpleTimer();
+				Optional<U> result = Optional.empty();
+				try {
+					
+						result = Optional.of(s.get());
+						try {
+							long elapsedNanos= (long)(jitterInNanos * r.nextDouble());
+							long millis = elapsedNanos/1000000;
+							int nanos = (int)(elapsedNanos - millis*1000000);
+							Thread.sleep(Math.max(0,millis),Math.max(0,nanos));
+						} catch (InterruptedException e) {
+							softener.throwSoftenedException(e);
+						}
+				} catch (ClosedQueueException e) {
+					if(result.isPresent())
+						throw new ClosedQueueException(result);
+					else
+						throw new ClosedQueueException();
+				}
+				return result.get();
+			};
+		};
+		return fromStream(queue.streamControl(getSubscription(), fn));
+	}
 	/**
 	 * Apply a fixed delay before emitting elements to the next phase of the Stream.
 	 * Note this doesn't neccessarily imply a fixed delay between element creation (although it may do).
@@ -249,7 +471,35 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param unit TimeUnit for emissions
 	 * @return Next Stage of the Stream
 	 */
-	FutureStream<U> fixedDelay(long time, TimeUnit unit) ;
+	default FutureStream<U> fixedDelay(long time, TimeUnit unit) {
+		Queue queue = toQueue();
+		Function<Supplier<U>, Supplier<U>> fn = s -> {
+			return () -> {
+				SimpleTimer timer = new SimpleTimer();
+				Optional<U> result = Optional.empty();
+				try {
+					
+						result = Optional.of(s.get());
+						try {
+							long elapsedNanos= unit.toNanos(time);
+							long millis = elapsedNanos/1000000;
+							int nanos = (int)(elapsedNanos - millis*1000000);
+							Thread.sleep(Math.max(0,millis),Math.max(0,nanos));
+							
+						} catch (InterruptedException e) {
+							softener.throwSoftenedException(e);
+						}
+				} catch (ClosedQueueException e) {
+					if(result.isPresent())
+						throw new ClosedQueueException(result);
+					else
+						throw new ClosedQueueException();
+				}
+				return result.get();
+			};
+		};
+		return fromStream(queue.streamControl(getSubscription(), fn));
+	}
 	/**
 	 * Allows clients to control the emission of data for the next phase of the Stream.
 	 * The user specified function can delay, drop, or change elements
@@ -258,7 +508,10 @@ public interface FutureStream<U> extends Seq<U> {
 	 *           This function should return a Supplier which returns the desired result for the next element (or just the next element).
 	 * @return Next stage in Stream
 	 */
-	FutureStream<U> control(Function<Supplier<U>, Supplier<U>> fn);
+	default FutureStream<U> control(Function<Supplier<U>, Supplier<U>> fn){
+		Queue queue = toQueue();
+		return fromStream(queue.streamControl(getSubscription(), fn));
+	}
 	
 	/**
 	 * Can be used to debounce (accept a single data point from a unit of time) data.
@@ -268,7 +521,12 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param unit Time unit for specified time
 	 * @return Next stage of stream, with only 1 element per specified time windows
 	 */
-	FutureStream<U> debounce(long time, TimeUnit unit);
+	default FutureStream<U> debounce(long time, TimeUnit unit) {
+		Queue queue = toQueue();
+		long timeNanos =  unit.toNanos(time);
+		Function<Supplier<U>, Supplier<U>> fn = new Debounce<>(timeNanos,1l,true);
+		return fromStream(queue.streamControl(getSubscription(), fn));
+	}
 	/**
 	 * Slow emissions down, emiting one element per specified time period
 	 * 
@@ -276,7 +534,37 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param unit Time unit for frequency period
 	 * @return Stream with emissions slowed down by specified emission frequency
 	 */
-	FutureStream<U> onePer(long time, TimeUnit unit) ;
+	default FutureStream<U> onePer(long time, TimeUnit unit) {
+		Queue queue = toQueue();
+		Function<Supplier<U>, Supplier<U>> fn = s -> {
+			
+			return () -> {
+				SimpleTimer timer=  new SimpleTimer();
+				Optional<U> result = Optional.empty();
+				try {
+					
+						
+						try {
+							long elapsedNanos= unit.toNanos(time)- timer.getElapsedNanoseconds();
+							long millis = elapsedNanos/1000000;
+							int nanos = (int)(elapsedNanos - millis*1000000);
+							Thread.sleep(Math.max(0,millis),Math.max(0,nanos));
+						} catch (InterruptedException e) {
+							softener.throwSoftenedException(e);
+						}
+						result = Optional.of(s.get());
+						
+				} catch (ClosedQueueException e) {
+					if(result.isPresent())
+						throw new ClosedQueueException(result);
+					else
+						throw new ClosedQueueException();
+				}
+				return result.get();
+			};
+		};
+		return fromStream(queue.streamControl(getSubscription(), fn));
+	}
 	/**
 	 * Allows x (specified number of) emissions with a time period before stopping emmissions until specified time has elapsed since last emission
 	 * 
@@ -285,7 +573,43 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param unit Frequency time unit
 	 * @return Stream with emissions slowed down by specified emission frequency
 	 */
-	FutureStream<U> xPer(int x,long time, TimeUnit unit);
+	default FutureStream<U> xPer(int x,long time, TimeUnit unit) {
+		Queue queue = toQueue();
+		Function<Supplier<U>, Supplier<U>> fn = s -> {
+			SimpleTimer[] timer=  {new SimpleTimer()};
+			final int[]count={0};
+			return () -> {
+				
+				Optional<U> result = Optional.empty();
+				try {
+					
+						
+						try {
+							if(count[0]==x){
+								long elapsedNanos= unit.toNanos(time)-timer[0].getElapsedNanoseconds();
+								long millis = elapsedNanos/1000000;
+								int nanos = (int)(elapsedNanos - millis*1000000);
+								Thread.sleep(Math.max(0,millis),nanos);
+								count[0]=0;
+								timer[0]= new SimpleTimer();
+							}
+						} catch (InterruptedException e) {
+							softener.throwSoftenedException(e);
+						}
+						result = Optional.of(s.get());
+						
+				} catch (ClosedQueueException e) {
+					if(result.isPresent())
+						throw new ClosedQueueException(result);
+					else
+						throw new ClosedQueueException();
+				}
+				return result.get();
+			};
+		};
+		return fromStream(queue.streamControl(getSubscription(), fn));
+	}
+
 	/**
 	 * Organise elements in a Stream into a Collections based on the time period they pass through this stage
 	 * Excludes Null values (neccessary for timeout handling)
@@ -294,7 +618,13 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param unit Time unit during which all elements should be collected
 	 * @return Stream of Lists
 	 */
-	FutureStream<Collection<U>> batchByTime(long time, TimeUnit unit);
+	default FutureStream<Collection<U>> batchByTime(long time, TimeUnit unit) {
+		Queue queue = toQueue();
+		Function<BiFunction<Long,TimeUnit,U>, Supplier<Collection<U>>> fn = new BatchByTime<U>(time,unit,
+				this.getSubscription(),queue,()->new ArrayList<>());
+		
+		return fromStream(queue.streamBatch(getSubscription(), fn)).filter(c->!((Collection)c).isEmpty());
+	}
 	/**
 	 * Organise elements in a Stream into a Collections based on the time period they pass through this stage
 	 * 
@@ -303,7 +633,11 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param factory Instantiates the collections used in the batching
 	 * @return Stream of collections
 	 */
-	FutureStream<Collection<U>> batchByTime(long time, TimeUnit unit,Supplier<Collection<U>> factory);
+	default FutureStream<Collection<U>> batchByTime(long time, TimeUnit unit,Supplier<Collection<U>> factory) {
+		Queue queue = toQueue();
+		Function<BiFunction<Long,TimeUnit,U>, Supplier<Collection<U>>> fn = new BatchByTime<U>(time,unit, this.getSubscription(),queue,factory);
+		return fromStream(queue.streamBatch(getSubscription(), fn)).filter(c->!((Collection)c).isEmpty());
+	}
 
 	/**
 	 * 
@@ -313,7 +647,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param s Stream to merge with
 	 * @return Stream of Tuples with the latest values from this stream
 	 */
-	<T> FutureStream<Tuple2<U, T>> withLatest(FutureStream<T> s);
+	default <T> FutureStream<Tuple2<U, T>> withLatest(FutureStream<T> s) {
+		return fromStream(FutureStreamFunctions.withLatest(this, s));
+	}
 	/**
 	 * Similar to zip and withLatest, except will always take the latest from either Stream (merged with last available from the other).
 	 * By contrast zip takes new / latest values from both Streams and withLatest will always take the latest from this Stream while 
@@ -322,7 +658,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param s Stream to merge with
 	 * @return  Stream of Tuples with the latest values from either stream
 	 */
-	<T> FutureStream<Tuple2<U, T>> combineLatest(FutureStream<T> s);
+	default <T> FutureStream<Tuple2<U, T>> combineLatest(FutureStream<T> s) {
+		return fromStream(FutureStreamFunctions.combineLatest(this, s));
+	}
 	/**
 	 * Return a Stream with the same values as this Stream, but with all values omitted until the provided stream starts emitting values.
 	 * Provided Stream ends the stream of values from this stream.
@@ -330,7 +668,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param s Stream that will start the emission of values from this stream
 	 * @return Next stage in the Stream but with all values skipped until the provided Stream starts emitting
 	 */
-	<T>  FutureStream<U> skipUntil(FutureStream<T> s);
+	default<T>  FutureStream<U> skipUntil(FutureStream<T> s) {
+		return fromStream(FutureStreamFunctions.skipUntil(this, s));
+	}
 	/**
 	 * Return a Stream with the same values, but will stop emitting values once the provided Stream starts to emit values.
 	 * e.g. if the provided Stream is asynchronously refreshing state from some remote store, this stream can proceed until
@@ -339,113 +679,17 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @param s Stream that will stop the emission of values from this stream
 	 * @return Next stage in the Stream but will only emit values until provided Stream starts emitting values
 	 */
-	<T>  FutureStream<U> takeUntil(FutureStream<T> s);
-
-	
-	
-	/*
-	 * @see
-	 * com.aol.simple.react.stream.traits.SimpleReactStream#retry(java.util.
-	 * function.Function)
-	 */
-	<R> FutureStream<R> retry(final Function<U, R> fn);
-
-	/*
-	 * @see
-	 * com.aol.simple.react.stream.traits.SimpleReactStream#allOf(java.util.
-	 * stream.Collector, java.util.function.Function)
-	
-	default <T, R> FutureStream<R> allOf(final Collector collector,
-			final Function<T, R> fn) {
-		return (FutureStream) SimpleReactStream.super.allOf(collector, fn);
+	default<T>  FutureStream<U> takeUntil(FutureStream<T> s) {
+		return fromStream(FutureStreamFunctions.takeUntil(this, s));
 	}
 
-	default <R> FutureStream<R> anyOf(final Function<U, R> fn) {
-		return (FutureStream) SimpleReactStream.super.anyOf(fn);
-	}
- */
-	/*
-	 * @see com.aol.simple.react.stream.traits.SimpleReactStream#
-	 * fromStreamCompletableFuture(java.util.stream.Stream)
-	 */
-	 <R> FutureStream<R> fromStreamOfFutures(
-			Stream<CompletableFuture<R>> stream);
-	/*
-	 * @see
-	 * com.aol.simple.react.stream.traits.SimpleReactStream#then(java.util.function
-	 * .Function)
-	 */
-	<R> FutureStream<R> then(final Function<U, R> fn) ;
-	
-
-	/*
-	 * @see
-	 * com.aol.simple.react.stream.traits.SimpleReactStream#fromStream(java.
-	 * util.stream.Stream)
-	 */
 	<R> FutureStream<R> fromStream(Stream<R> stream);
-	/**
-	 * Perform a flatMap operation where the CompletableFuture type returned is flattened from the resulting Stream
-	 * If in async mode this operation is performed asyncrhonously
-	 * If in sync mode this operation is performed synchronously
-	 * 
-	 * <pre>
-	 * {@code 
-	 * assertThat( new LazyReact()
-										.of(1,2,3)
-										.flatMapCompletableFuture(i->CompletableFuture.completedFuture(i))
-										.block(),equalTo(Arrays.asList(1,2,3)));
-	 * }</pre>
-	 *
-	 * In this example the result of the flatMapCompletableFuture is 'flattened' to the raw integer values
-	 * 
-	 * 
-	 * @param flatFn flatMap function
-	 * @return Flatten Stream with flatFn applied
-	 */
-	 <R> FutureStream<R> flatMapCompletableFuture(
-			Function<U, CompletableFuture<R>> flatFn) ;
-	/**
-	 * Perform a flatMap operation where the CompletableFuture type returned is flattened from the resulting Stream
-	 * This operation is performed synchronously
-	 * 
-	 * <pre>
-	 * {@code 
-	 * assertThat( new LazyReact()
-										.of(1,2,3)
-										.flatMapCompletableFutureSync(i->CompletableFuture.completedFuture(i))
-										.block(),equalTo(Arrays.asList(1,2,3)));
-	 * }
-	 * </pre>
-	 * In this example the result of the flatMapCompletableFuture is 'flattened' to the raw integer values
-	 * 
-	 * 
-	 * @param flatFn flatMap function
-	 * @return Flatten Stream with flatFn applied
-	 */
-	 <R> FutureStream<R> flatMapCompletableFutureSync(
-			Function<U, CompletableFuture<R>> flatFn);
+	 <R> FutureStream<R> flatMap(
+				Function<? super U, ? extends Stream<? extends R>> flatFn);
+	
+	 FutureStream<U> filter(final Predicate<? super U> p) ;
+	 FutureStream<U> peek(final Consumer<? super U> consumer);
 
-	/*
-	 * @see org.jooq.lambda.Seq#flatMap(java.util.function.Function)
-	 */
-	@Override
-	<R> FutureStream<R> flatMap(
-			Function<? super U, ? extends Stream<? extends R>> flatFn);
-	
-	
-	
-	
-	/*
-	 * @see org.jooq.lambda.Seq#peek(java.util.function.Consumer)
-	 */
-	@Override
-	FutureStream<U> peek(final Consumer<? super U> consumer);
-
-	/*
-	 * @see org.jooq.lambda.Seq#filter(java.util.function.Predicate)
-	 */
-	FutureStream<U> filter(final Predicate<? super U> p) ;
 
 	/**
 	 * Stream supporting methods
@@ -460,7 +704,11 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#forEach(java.util.function.Consumer)
 	 */
 	@Override
-	void forEach(Consumer<? super U> action);
+	default void forEach(Consumer<? super U> action) {
+		toQueue().stream(getSubscription()).forEach((Consumer) action);
+
+	}
+
 	
 	/* 
 	 * Sequentially iterate through the LazyFutureStream
@@ -470,7 +718,10 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#forEachOrdered(java.util.function.Consumer)
 	 */
 	@Override
-	void forEachOrdered(Consumer<? super U> action);
+	default void forEachOrdered(Consumer<? super U> action) {
+		toQueue().stream(getSubscription()).forEachOrdered((Consumer) action);
+
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -478,7 +729,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#toArray()
 	 */
 	@Override
-	Object[] toArray();
+	default Object[] toArray() {
+		return toQueue().stream(getSubscription()).toArray();
+	}
 	
 	/**
 	 * Create a sliding view over this Stream
@@ -499,7 +752,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 *            Size of sliding window
 	 * @return Stream with sliding view over data in this stream
 	 */
-	FutureStream<List<U>> sliding(int size);
+	default FutureStream<List<U>> sliding(int size){
+		return this.fromStream(SlidingWindow.sliding(this,size, 1));
+	}
 	/**
 	 * Create a sliding view over this Stream
 	 * 
@@ -519,17 +774,24 @@ public interface FutureStream<U> extends Seq<U> {
 	 *            Size of sliding window
 	 * @return Stream with sliding view over data in this stream
 	 */
-	FutureStream<List<U>> sliding(int size, int increment);
+	default FutureStream<List<U>> sliding(int size, int increment){
+		return this.fromStream(SlidingWindow.sliding(this,size, increment));
+	}
 	
 	/**
 	 * @return An extensive set of asyncrhonous terminal operations
 	 */
-	FutureOps<U> futureOperations();
+	default FutureOps<U> futureOperations(){
+		return new FutureOps<>(this.getTaskExecutor(),this);
+	}
+	
 	/**
 	 * @param executor to execute terminal ops asynchronously on
 	 * @return An extensive set of asyncrhonous terminal operations
 	 */
-	FutureOps<U> futureOperations(Executor executor);
+	default FutureOps<U> futureOperations(Executor executor){
+		return new FutureOps<>(executor,this);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -537,7 +799,10 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#toArray(java.util.function.IntFunction)
 	 */
 	@Override
-	<A> A[] toArray(IntFunction<A[]> generator);
+	default <A> A[] toArray(IntFunction<A[]> generator) {
+		return toQueue().stream(getSubscription()).toArray(generator);
+	}
+
 	/*
 	 * (non-Javadoc)
 	 * 
@@ -545,7 +810,11 @@ public interface FutureStream<U> extends Seq<U> {
 	 * java.util.function.BinaryOperator)
 	 */
 	@Override
-	U reduce(U identity, BinaryOperator<U> accumulator);
+	default U reduce(U identity, BinaryOperator<U> accumulator) {
+
+		return (U) toQueue().stream(getSubscription()).reduce(identity,
+				accumulator);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -553,7 +822,10 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#reduce(java.util.function.BinaryOperator)
 	 */
 	@Override
-	Optional<U> reduce(BinaryOperator<U> accumulator);
+	default Optional<U> reduce(BinaryOperator<U> accumulator) {
+		return toQueue().stream(getSubscription()).reduce(accumulator);
+	}
+
 	/*
 	 * (non-Javadoc)
 	 * 
@@ -561,8 +833,12 @@ public interface FutureStream<U> extends Seq<U> {
 	 * java.util.function.BiConsumer, java.util.function.BiConsumer)
 	 */
 	@Override
-	 <R> R collect(Supplier<R> supplier,
-			BiConsumer<R, ? super U> accumulator, BiConsumer<R, R> combiner);
+	default <R> R collect(Supplier<R> supplier,
+			BiConsumer<R, ? super U> accumulator, BiConsumer<R, R> combiner) {
+
+		return (R) toQueue().stream(getSubscription()).collect(supplier,
+				accumulator, combiner);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -570,7 +846,10 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#min(java.util.Comparator)
 	 */
 	@Override
-	Optional<U> min(Comparator<? super U> comparator);
+	default Optional<U> min(Comparator<? super U> comparator) {
+
+		return toQueue().stream(getSubscription()).min(comparator);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -578,7 +857,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#max(java.util.Comparator)
 	 */
 	@Override
-	Optional<U> max(Comparator<? super U> comparator);
+	default Optional<U> max(Comparator<? super U> comparator) {
+		return toQueue().stream(getSubscription()).max(comparator);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -586,7 +867,10 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#count()
 	 */
 	@Override
-	long count() ;
+	long count();/** {
+
+		return getLastActive().stream().count();
+	}**/
 
 	/*
 	 * (non-Javadoc)
@@ -594,14 +878,19 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#anyMatch(java.util.function.Predicate)
 	 */
 	@Override
-	boolean anyMatch(Predicate<? super U> predicate);
+	default boolean anyMatch(Predicate<? super U> predicate) {
+		return toQueue().stream(getSubscription()).anyMatch(predicate);
+	}
+
 	/*
 	 * (non-Javadoc)
 	 * 
 	 * @see java.util.stream.Stream#allMatch(java.util.function.Predicate)
 	 */
 	@Override
-	boolean allMatch(Predicate<? super U> predicate) ;
+	default boolean allMatch(Predicate<? super U> predicate) {
+		return toQueue().stream(getSubscription()).allMatch(predicate);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -609,7 +898,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#noneMatch(java.util.function.Predicate)
 	 */
 	@Override
-	boolean noneMatch(Predicate<? super U> predicate);
+	default boolean noneMatch(Predicate<? super U> predicate) {
+		return toQueue().stream(getSubscription()).noneMatch(predicate);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -617,7 +908,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#findFirst()
 	 */
 	@Override
-	Optional<U> findFirst() ;
+	default Optional<U> findFirst() {
+		return toQueue().stream(getSubscription()).findFirst();
+	}
 	/**
 	 * <pre>
 	 * {@code 
@@ -631,7 +924,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @return the firstValue in this stream - same as findFirst but doesn't 
 	 * 				return an Optional. Will throw an exception if empty
 	 */
-	U firstValue();
+	default U firstValue() {
+		return toQueue().stream(getSubscription()).findFirst().get();
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -639,7 +934,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.Stream#findAny()
 	 */
 	@Override
-	Optional<U> findAny();
+	default Optional<U> findAny() {
+		return toQueue().stream(getSubscription()).findAny();
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -648,8 +945,12 @@ public interface FutureStream<U> extends Seq<U> {
 	 * java.util.function.BiFunction, java.util.function.BinaryOperator)
 	 */
 	@Override
-	<R> R reduce(R identity, BiFunction<R, ? super U, R> accumulator,
-			BinaryOperator<R> combiner);
+	default <R> R reduce(R identity, BiFunction<R, ? super U, R> accumulator,
+			BinaryOperator<R> combiner) {
+
+		return toQueue().stream(getSubscription()).reduce(identity,
+				accumulator, combiner);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -657,7 +958,16 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see java.util.stream.BaseStream#iterator()
 	 */
 	@Override
-	CloseableIterator<U> iterator();
+	default CloseableIterator<U> iterator() {
+
+		Queue<U> q = toQueue();
+		if (getSubscription().closed())
+			return new CloseableIterator<>(Arrays.<U> asList().iterator(),
+					getSubscription(),null);
+		
+		return new CloseableIterator<>(q.stream(getSubscription())
+				.iterator(), getSubscription(),q);
+	}
 	
  
 	
@@ -668,8 +978,31 @@ public interface FutureStream<U> extends Seq<U> {
 	  *	@return
 	  * @see org.jooq.lambda.Seq#reverse()
 	  */
-	FutureStream<U> reverse();
-	Iterator<U> reversedIterator();
+	default FutureStream<U> reverse() {
+	        return fromStream(Seq.seq(reversedIterator()));
+	   }
+	default Iterator<U> reversedIterator(){
+		Queue<U> q = toQueue();
+		if (getSubscription().closed())
+			return new CloseableIterator<>(Arrays.<U> asList().iterator(),
+					getSubscription(),null);
+		List l=  q.stream(getSubscription()).toList();
+		ListIterator iterator = l.listIterator(l.size());
+		return new CloseableIterator<>(new Iterator(){
+
+			@Override
+			public boolean hasNext() {
+				return iterator.hasPrevious();
+			}
+
+			@Override
+			public Object next() {
+				return iterator.previous();
+			}
+			
+		}, getSubscription(),q);
+		
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -677,14 +1010,19 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#spliterator()
 	 */
 	@Override
-	Spliterator<U> spliterator();
+	default Spliterator<U> spliterator() {
+		return toQueue().stream(getSubscription()).spliterator();
+	}
+
 	/*
 	 * (non-Javadoc)
 	 * 
 	 * @see java.util.stream.BaseStream#isParallel()
 	 */
 	@Override
-	boolean isParallel();
+	default boolean isParallel(){
+		return true;
+	}
 
 	/*
 	 * Creates a sequential instance by populating an async Queue from the
@@ -698,12 +1036,20 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see com.aol.simple.react.stream.traits.FutureStream#sequential()
 	 */
 	@Override
-	Seq<U> sequential();
+	 Seq<U> sequential();/** {
+		Queue q = toQueue();
+		q.fromStream(getLastActive().stream().map(it -> it.join()));
+		q.close();
+		return q.stream();
+	}**/
 
 
 
 	@Override
-	Stream<U> stream();
+	default Stream<U> stream() {
+		return toQueue().stream(getSubscription());
+	}
+
 	/*
 	 * Seq supporting methods
 	 */
@@ -723,7 +1069,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 *
 	 * @see #intersperse(Stream, Object)
 	 */
-	FutureStream<U> intersperse(U value);
+	default FutureStream<U> intersperse(U value) {
+		return intersperse(this, value);
+	}
 
 	/*
 	 * 
@@ -735,12 +1083,52 @@ public interface FutureStream<U> extends Seq<U> {
 	 */
 
 	@Override
-	<U> FutureStream<U> cast(Class<U> type);
+	default <U> FutureStream<U> cast(Class<U> type) {
+		return (FutureStream<U>) cast(this, type);
+	}
 
 	@Override
-	<U> FutureStream<U> ofType(Class<U> type);
+	default <U> FutureStream<U> ofType(Class<U> type) {
+		return (FutureStream<U>) ofType(this, type);
+	}
 
-	
+	/**
+	 * Keep only those elements in a stream that are of a given type.
+	 * 
+	 * 
+	 * // (1, 2, 3) EagerFutureStream.of(1, "a", 2, "b",
+	 * 3).ofType(Integer.class)
+	 * 
+	 */
+	@SuppressWarnings("unchecked")
+	static <T, U> FutureStream<U> ofType(FutureStream<T> stream, Class<U> type) {
+		return stream.filter(type::isInstance).map(t -> (U) t);
+	}
+
+	/**
+	 * Cast all elements in a stream to a given type, possibly throwing a
+	 * {@link ClassCastException}.
+	 * 
+	 * 
+	 * // ClassCastException LazyFutureStream.of(1, "a", 2, "b",
+	 * 3).cast(Integer.class)
+	 * 
+	 */
+	static <T, U> FutureStream<U> cast(FutureStream<T> stream, Class<U> type) {
+		return stream.map(type::cast);
+	}
+
+	/**
+	 * Returns a stream with a given value interspersed between any two values
+	 * of this stream.
+	 * 
+	 * 
+	 * // (1, 0, 2, 0, 3, 0, 4) Seq.of(1, 2, 3, 4).intersperse(0)
+	 * 
+	 */
+	static <T> FutureStream<T> intersperse(FutureStream<T> stream, T value) {
+		return stream.flatMap(t -> Stream.of(value, t).skip(1));
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -748,7 +1136,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#unordered()
 	 */
 	@Override
-	FutureStream<U> unordered() ;
+	default FutureStream<U> unordered() {
+		return this;
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -756,7 +1146,11 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#onClose(java.lang.Runnable)
 	 */
 	@Override
-	 Seq<U> onClose(Runnable closeHandler);
+	Seq<U> onClose(Runnable closeHandler);/** {
+
+		return Seq.seq(getLastActive().stream().onClose(closeHandler)
+				.map(it -> (U) it.join()));
+	}**/
 
 	/*
 	 * (non-Javadoc)
@@ -764,7 +1158,10 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#close()
 	 */
 	@Override
-	 void close() ;
+	void close();/** {
+		getLastActive().stream().close();
+
+	}**/
 
 	/*
 	 * (non-Javadoc)
@@ -772,7 +1169,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#map(java.util.function.Function)
 	 */
 	@Override
-	<R> FutureStream<R> map(Function<? super U, ? extends R> mapper);
+	default <R> FutureStream<R> map(Function<? super U, ? extends R> mapper) {
+		return (FutureStream)then((Function) mapper);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -780,7 +1179,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#mapToInt(java.util.function.ToIntFunction)
 	 */
 	@Override
-	IntStream mapToInt(ToIntFunction<? super U> mapper);
+	default IntStream mapToInt(ToIntFunction<? super U> mapper) {
+		return toQueue().stream(getSubscription()).mapToInt(mapper);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -788,7 +1189,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#mapToLong(java.util.function.ToLongFunction)
 	 */
 	@Override
-	LongStream mapToLong(ToLongFunction<? super U> mapper);
+	default LongStream mapToLong(ToLongFunction<? super U> mapper) {
+		return toQueue().stream(getSubscription()).mapToLong(mapper);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -796,7 +1199,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#mapToDouble(java.util.function.ToDoubleFunction)
 	 */
 	@Override
-	DoubleStream mapToDouble(ToDoubleFunction<? super U> mapper) ;
+	default DoubleStream mapToDouble(ToDoubleFunction<? super U> mapper) {
+		return toQueue().stream(getSubscription()).mapToDouble(mapper);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -804,24 +1209,32 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#flatMapToInt(java.util.function.Function)
 	 */
 	@Override
-	IntStream flatMapToInt(
-			Function<? super U, ? extends IntStream> mapper) ;
+	default IntStream flatMapToInt(
+			Function<? super U, ? extends IntStream> mapper) {
+		return toQueue().stream(getSubscription()).flatMapToInt(mapper);
+	}
+
 	/*
 	 * (non-Javadoc)
 	 * 
 	 * @see org.jooq.lambda.Seq#flatMapToLong(java.util.function.Function)
 	 */
 	@Override
-	LongStream flatMapToLong(
-			Function<? super U, ? extends LongStream> mapper);
+	default LongStream flatMapToLong(
+			Function<? super U, ? extends LongStream> mapper) {
+		return toQueue().stream(getSubscription()).flatMapToLong(mapper);
+	}
+
 	/*
 	 * (non-Javadoc)
 	 * 
 	 * @see org.jooq.lambda.Seq#flatMapToDouble(java.util.function.Function)
 	 */
 	@Override
-	DoubleStream flatMapToDouble(
-			Function<? super U, ? extends DoubleStream> mapper) ;
+	default DoubleStream flatMapToDouble(
+			Function<? super U, ? extends DoubleStream> mapper) {
+		return toQueue().stream(getSubscription()).flatMapToDouble(mapper);
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -829,7 +1242,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#sorted()
 	 */
 	@Override
-	Seq<U> sorted();
+	default Seq<U> sorted() {
+		return toQueue().stream(getSubscription()).sorted();
+	}
 
 	/*
 	 * (non-Javadoc)
@@ -837,7 +1252,9 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see org.jooq.lambda.Seq#sorted(java.util.Comparator)
 	 */
 	@Override
-	Seq<U> sorted(Comparator<? super U> comparator);
+	default Seq<U> sorted(Comparator<? super U> comparator) {
+		return toQueue().stream(getSubscription()).sorted(comparator);
+	}
 
 	/**
 	 * Give a function access to the current stage of a SimpleReact Stream
@@ -846,7 +1263,13 @@ public interface FutureStream<U> extends Seq<U> {
 	 *            Consumer that will recieve current stage
 	 * @return Self (current stage)
 	 */
-	FutureStream<U> self(Consumer<FutureStream<U>> consumer);
+	default FutureStream<U> self(Consumer<FutureStream<U>> consumer) {
+		return (FutureStream<U>) then((t) -> {
+			consumer.accept(this);
+			return (U) t;
+		});
+
+	}
 
 	/**
 	 * Returns a limited interval from a given Stream.
@@ -858,37 +1281,31 @@ public interface FutureStream<U> extends Seq<U> {
 	 * @see #slice(Stream, long, long)
 	 */
 	@Override
-	Seq<U> slice(long from, long to) ;
+	default Seq<U> slice(long from, long to) {
+		return Seq.super.slice(from, to);
+	}
 
-	
-
-	
-
-	
-	/* 
-	 * Execute subsequent stages on the completing thread (until async called)
-	 * 10X faster than async execution.
-	 * Use async for blocking IO or distributing work across threads or cores.
-	 * Switch to sync for non-blocking tasks when desired thread utlisation reached
+	/**
+	 * Returns a limited interval from a given Stream.
 	 * 
-	 *	@return Version of FutureStream that will use sync CompletableFuture methods
-	 * @see com.aol.simple.react.stream.traits.SimpleReactStream#sync()
+	 * 
+	 * // (4, 5) EagerFutureStream.of(1, 2, 3, 4, 5, 6).slice(3, 5)
+	 * 
 	 */
-	FutureStream<U> sync();
+	static <T> Seq<T> slice(FutureStream<T> stream, long from, long to) {
+		long f = Math.max(from, 0);
+		long t = Math.max(to - f, 0);
+
+		return stream.skip(f).limit(t);
+	}
+
 	
 
-	/* 
-	 * Execute subsequent stages by submission to an Executor for async execution
-	 * 10X slower than sync execution.
-	 * Use async for blocking IO or distributing work across threads or cores.
-	 * Switch to sync for non-blocking tasks when desired thread utlisation reached
-	 *
-	 * 
-	 *	@return Version of FutureStream that will use async CompletableFuture methods
-	 * @see com.aol.simple.react.stream.traits.SimpleReactStream#async()
-	 */
-	FutureStream<U> async();
-	Queue<U> toQueue();
-	Continueable getSubscription();
+	
+	
+
+	
+	 Continueable getSubscription();
+	 Executor getTaskExecutor();
 	
 }
