+++ src/main/java/com/aol/simple/react/stream/traits/SimpleReactStreamInterface.java
@@ -536,7 +536,7 @@ public interface SimpleReactStreamInterface<U>{
 	 *
 	 */
 	SimpleReactStreamInterface<U> async();
-	Queue toQueue();
+	Queue<U> toQueue();
